refactor(editor): table affordance 모델을 분리한다
