+++ front/src/components/editor/BlockEditorEngine.tsx
@@ -88,6 +88,18 @@ import {
   createWriterEditorExtensions,
   TABLE_CONTEXT_BLOCKED_INSERT_IDS,
 } from "./writerEditorPreset"
+import {
+  type TableAffordanceGeometry,
+  type TableAffordanceVisibility,
+  INITIAL_TABLE_AFFORDANCE_GEOMETRY,
+  INITIAL_TABLE_AFFORDANCE_VISIBILITY,
+  TABLE_ADD_BAR_VIEWPORT_PADDING_PX,
+  TABLE_CELL_MENU_BUTTON_SIZE_PX,
+  TABLE_EDGE_ADD_BUTTON_SIZE_PX,
+  clampViewportPosition,
+  intersectsViewportBounds,
+  resolveDesktopTableRailLayout,
+} from "./tableAffordanceModel"
 
 type RuntimeGuardWindow = Window & {
   __AQ_RUNTIME_GUARD_ENABLED__?: boolean
@@ -255,56 +267,6 @@ type FloatingBubbleState = {
   top: number
 }
 
-type TableOverlayAnchor = {
-  left: number
-  top: number
-}
-
-type TableAffordanceGeometry = {
-  left: number
-  top: number
-  tableLeft: number
-  tableTop: number
-  tableRight: number
-  tableBottom: number
-  width: number
-  height: number
-  surfaceLeft: number
-  surfaceTop: number
-  surfaceWidth: number
-  surfaceHeight: number
-  cellLeft: number
-  cellTop: number
-  cellWidth: number
-  cellHeight: number
-  rowIndex: number
-  rowTop: number
-  rowHeight: number
-  columnLeft: number
-  columnWidth: number
-  columnIndex: number
-  rowHandleAnchor: TableOverlayAnchor
-  columnHandleAnchor: TableOverlayAnchor
-  rowAddAnchor: TableOverlayAnchor
-  columnAddAnchor: TableOverlayAnchor
-  cornerAnchor: TableOverlayAnchor
-  cellMenuAnchor: TableOverlayAnchor
-  columnSegments: Array<{
-    left: number
-    width: number
-  }>
-}
-
-type TableAffordanceVisibility = {
-  visible: boolean
-  showColumnRail: boolean
-  showRowRail: boolean
-  showColumnAddBar: boolean
-  showRowAddBar: boolean
-  showCornerControls: boolean
-  showCellMenu: boolean
-}
-
 type TableMenuKind = "row" | "column" | "table" | "cell"
 
 type TableMenuState =
@@ -465,48 +427,6 @@ type TableAxisReorderIndicatorState = {
   top: number
   width: number
   height: number
-}
-
-const INITIAL_TABLE_AFFORDANCE_GEOMETRY: TableAffordanceGeometry = {
-  left: 0,
-  top: 0,
-  tableLeft: 0,
-  tableTop: 0,
-  tableRight: 0,
-  tableBottom: 0,
-  width: 0,
-  height: 0,
-  surfaceLeft: 0,
-  surfaceTop: 0,
-  surfaceWidth: 0,
-  surfaceHeight: 0,
-  cellLeft: 0,
-  cellTop: 0,
-  cellWidth: 0,
-  cellHeight: 0,
-  rowIndex: 0,
-  rowTop: 0,
-  rowHeight: 0,
-  columnLeft: 0,
-  columnWidth: 0,
-  columnIndex: 0,
-  rowHandleAnchor: { left: 0, top: 0 },
-  columnHandleAnchor: { left: 0, top: 0 },
-  rowAddAnchor: { left: 0, top: 0 },
-  columnAddAnchor: { left: 0, top: 0 },
-  cornerAnchor: { left: 0, top: 0 },
-  cellMenuAnchor: { left: 0, top: 0 },
-  columnSegments: [],
-}
-
-const INITIAL_TABLE_AFFORDANCE_VISIBILITY: TableAffordanceVisibility = {
-  visible: false,
-  showColumnRail: false,
-  showRowRail: false,
-  showColumnAddBar: false,
-  showRowAddBar: false,
-  showCornerControls: false,
-  showCellMenu: false,
 }
 
 type BlockSelectionPointerEventLike = {
@@ -965,7 +885,6 @@ const TABLE_ROW_GRIP_BUTTON_WIDTH_PX = 16
 const TABLE_ROW_GRIP_BUTTON_HEIGHT_PX = 26
 const TABLE_AXIS_GRIP_EDGE_INSET_PX = 0
 const TABLE_ADD_BAR_THICKNESS_PX = 28
-const TABLE_ADD_BAR_VIEWPORT_PADDING_PX = 8
 const TABLE_AXIS_RAIL_EDGE_HOTZONE_PX = 18
 const TABLE_TRAILING_ADD_EDGE_HOTZONE_PX = 18
 const TABLE_OVERFLOW_COACHMARK_ESTIMATED_WIDTH_PX = 244
@@ -1134,11 +1053,9 @@ const countShrinkableRenderedTableAxisAtEnd = (tableElement: HTMLTableElement | 
   return shrinkableColumns
 }
 const TABLE_EDGE_HANDLE_INSET_PX = 6
-const TABLE_EDGE_ADD_BUTTON_SIZE_PX = 24
 const TABLE_CORNER_CLUSTER_GAP_PX = 6
 const TABLE_CORNER_CLUSTER_WIDTH_PX =
   TABLE_CORNER_BUTTON_SIZE_PX * 2 + TABLE_CORNER_CLUSTER_GAP_PX
-const TABLE_CELL_MENU_BUTTON_SIZE_PX = 22
 const TABLE_QUICK_RAIL_HIDE_DELAY_MS = 120
 const TABLE_MENU_EDGE_PADDING_PX = 16
 const TABLE_MENU_ESTIMATED_WIDTH_PX = 272
@@ -1162,95 +1079,6 @@ const TABLE_CELL_COLOR_PRESETS = [
   { label: "라일락", value: "#ddd6fe" },
   { label: "회색", value: "#e2e8f0" },
 ] as const
-
-const clampViewportPosition = (
-  value: number,
-  edgePadding: number,
-  viewportSize: number,
-  itemSize: number
-) => {
-  const max = Math.max(edgePadding, viewportSize - itemSize - edgePadding)
-  return Math.min(Math.max(value, edgePadding), max)
-}
-
-const intersectsViewportBounds = (
-  rect: DOMRect | null | undefined,
-  edgePadding = TABLE_ADD_BAR_VIEWPORT_PADDING_PX
-) => {
-  if (!rect) return false
-  if (typeof window === "undefined") return true
-  return (
-    rect.right >= edgePadding &&
-    rect.bottom >= edgePadding &&
-    rect.left <= window.innerWidth - edgePadding &&
-    rect.top <= window.innerHeight - edgePadding
-  )
-}
-
-const resolveDesktopTableRailLayout = (
-  state: TableAffordanceGeometry
-) => {
-  const columnGripTop = Math.round(state.columnHandleAnchor.top)
-  const cornerTop = Math.round(state.cornerAnchor.top)
-  const cornerLeft = Math.round(state.cornerAnchor.left)
-  const columnGripLeft = Math.round(state.columnHandleAnchor.left)
-  const columnAddBarLeft =
-    typeof window === "undefined"
-      ? Math.round(state.columnAddAnchor.left)
-      : clampViewportPosition(
-          Math.round(state.columnAddAnchor.left),
-          TABLE_ADD_BAR_VIEWPORT_PADDING_PX,
-          window.innerWidth,
-          TABLE_EDGE_ADD_BUTTON_SIZE_PX
-        )
-  const columnAddBarTop = Math.round(state.columnAddAnchor.top)
-  const rowGripTop = Math.round(state.rowHandleAnchor.top)
-  const rowGripLeft = Math.round(state.rowHandleAnchor.left)
-  const rowAddBarLeft = Math.round(state.rowAddAnchor.left)
-  const rowAddBarTop =
-    typeof window === "undefined"
-      ? Math.round(state.rowAddAnchor.top)
-      : clampViewportPosition(
-          Math.round(state.rowAddAnchor.top),
-          TABLE_ADD_BAR_VIEWPORT_PADDING_PX,
-          window.innerHeight,
-          TABLE_EDGE_ADD_BUTTON_SIZE_PX
-        )
-
-  const cellMenuLeft =
-    typeof window === "undefined"
-      ? Math.round(state.cellMenuAnchor.left)
-      : clampViewportPosition(
-          Math.round(state.cellMenuAnchor.left),
-          TABLE_ADD_BAR_VIEWPORT_PADDING_PX,
-          window.innerWidth,
-          TABLE_CELL_MENU_BUTTON_SIZE_PX
-        )
-  const cellMenuTop =
-    typeof window === "undefined"
-      ? Math.round(state.cellMenuAnchor.top)
-      : clampViewportPosition(
-          Math.round(state.cellMenuAnchor.top),
-          TABLE_ADD_BAR_VIEWPORT_PADDING_PX,
-          window.innerHeight,
-          TABLE_CELL_MENU_BUTTON_SIZE_PX
-        )
-
-  return {
-    cornerLeft,
-    cornerTop,
-    columnGripLeft,
-    columnGripTop,
-    columnAddBarLeft,
-    columnAddBarTop,
-    rowGripLeft,
-    rowGripTop,
-    rowAddBarLeft,
-    rowAddBarTop,
-    cellMenuLeft,
-    cellMenuTop,
-  }
-}
 
 const blockHasVisibleContent = (node?: BlockEditorDoc | null): boolean => {
   if (!node) return false
